Restructuring into a more lib-like format for easier integration
