+++ docker/Dockerfile
@@ -42,3 +42,4 @@ RUN pip install pyserial \
 ##### 
 RUN pip install torch torchvision
 RUN pip install opencv-python
+RUN pip install torchdiffeq pandas matplotlib
