+++ lecontrol/configuration/configs.py
@@ -1,6 +1,6 @@
 BAUDRATE = 1000000
-PORT_LEADER = '/dev/ttyACM1'
-PORT_FOLLOWER = '/dev/ttyACM0'
+PORT_LEADER = '/dev/ttyACM0'
+PORT_FOLLOWER = '/dev/ttyACM1'
 PROTOCOL_END = 0
 
 RATE = 100
@@ -11,7 +11,7 @@ CAMERAS = [
         'width': 640,
         'height': 480,
         'fps': 30,
-        'camera_id': 8,
+        'camera_id': 0,
     },
     # {
     #     'name': 'camera_2',
